+++ lib/link_preview/parallel_helper.ex
@@ -1,4 +1,4 @@
-defmodule LinkPreviewGenerator.ParallelHelper do
+defmodule LinkPreview.ParallelHelper do
   @moduledoc false
 
   def map(collection, fun) do