+++ test/link_preview/processor_test.exs
@@ -1,17 +1,17 @@
-defmodule LinkPreviewGenerator.ProcessorTest do
+defmodule LinkPreview.ProcessorTest do
   use ExUnit.Case
 
   @http "http://localhost:#{Application.get_env(:httparrot, :http_port)}"
 
   describe "call" do
     test "when url leads to image" do
-      %LinkPreviewGenerator.Page{
+      %LinkPreview.Page{
         title: nil,
         description: nil,
         original_url: original_url,
         website_url: website_url,
         images: images
-      } = LinkPreviewGenerator.Processor.call(@http <> "/image")
+      } = LinkPreview.Processor.call(@http <> "/image")
 
       assert original_url
       assert website_url